Move tests to src, use the new mocking helper, fix tests
